Add ttt_flash (#97)
* Add ttt_flash

Adds an upgrade where anyone who tases a t will be flashed for a given period of time.

* Subtle Optimisation

Should improve the performance of both invis and flash as it exits the code earlier depending on the victim.

* Some fixes

Fixed compile issues with bloodlust and corrected some missing parrameters.

* Fixed up the parameters

Adjusted the timings to the sufficient length in miliseconds. Also corrected the flags settings.