Create Base Special Days API
Created TTT Special Days is a main / API plugin that will  allow for the management of multiple special days via calling them with a generalised forward which passed the index for which special day to activate.
Created to Forwards one for calling to start a day and also another to stop them though this is only for debugging really etc.