Further Work on KotH and small changes to SD
Fix all typos in TTT Special Days.
Renamed the defines for TTT KotH and BHop for easier reading.
Added new stock to generics for setting convar bools. Will add further ones for ints etc.
Added a command to add hills to maps for KotH as well as created the config file and manager for it.
Added RayTrace to player method map. Is used in Placing hills but will also be needed for Teleport when general is ported.