TTT Ranks Optimisation. (#138)
* TTT Ranks Optimisation.

Used an array to store all players ranks instead of fetching it each time.
Renamed sm_rank to sm_rankcheck to avoid incompatibility.
Renamed sm_ranktest to sm_ranktemp to better reflect the functionality.
In-lined most internal Function Calls.

* Add Extra Safety Checks.

Add rank resets on MapChange, if Player.ValidClient returns true the rank is set normally else it is reset to 0, also this loops through all player index's not just current ones.

* Should fix Compile Issues.

Cant check atm as going home for the weekend but that should be out culprit.
Is the way it had had to be done throughout the rest of the code not sure why, may be to do with using natives.

* mini changes

* fix player viewing

* Fix Compile Issues. Changed to using LoopClients.

Reorganised some include orders.
Fix some more formatting.
Add checks into sm_refreshranks that where done by LoopValidClients.

* Fix Rank Ordering.

Senator is now again the highest rank. Though must be set in the config file still for now less another solution is adapted too or improved.
Could just check for the user z flag if they werent picked up in the config.