Add TTT Staff & Player Commands (#137)
* Add TTT Staff & Player Commands

Removed duplicate Player.RankName as we have Player.GetRankName which has more uses.
Added Teleport to Staff Commands.
Added Staff List, and team commands to Player Commands.

* Fix Indentation.

Replaced a tab with 4 spaces.

* Start of sm_rank.

Needs a better implementation of the sql query to also get the count and player position. Not sure if this can be done in this way for an async. Need to look into it more.

* Finalise SQL query for sm_rank

Added all needed functionality for sm_rank
Changed the sql query to return all rows with karma greater than that of the player and too use a dummy row to store the total player count.

* Add sm_playtime

Added all functionality in the for sm_playtime in the same style as sm_rank.

* Fix and Optimise sm_rank and sm_playtime

Change to using passed userID's for getting the client who called the commands, also remembered to pass client to CPrtintToChatAll in sm_playtime.
Changed the SQL query for sm_rank to be more optimised by only returning a single row.
Added null check in sm_rank to check whether the play is in first place as this takes place on nextKarma.
Renamed any data to any userID in both database callbacks to better reflect the usage.

* Correct Typos and Spelling

Renamed staffIndexs too staffIndexes.
Shortened and corrected playtime outputs.
Corrected some indentation.

* Add more commands sm_bantimes and sm_forcespec

Altered ban times from the original in TTT Popey to follow the current colouring and styling patterns.

* Add sm_give

Fix missing '}' in prints for sm_playtime.
Also added missing returns in both TTT Player Commands and TTT Staff Commands.

* Add sm_reload and fix Access checks.

Was using Player.Access instead of !Player.Access.
Changed print formatting.

* Rename TTT General New to TTT General