[Skills] Rewrite (#106)
* ReWrite ReducedFalling

* ReWrite Scavenger

Also improved some of the code to maybe function better.

* Changes to Scavenger

Forgot to readd a variable to the if and swapped valid client for the property version in player method map

* ReWrite Binoculars

I am not sure why initialFov still exist as it is not used in the code except storing it, also lastTime doesnt affect the code too much but I'll leave them both in for now.

* Fix Binoculars

Fixed max level and an if statement.

* ReWrite Vampire

Corrected some code layout and scopes for better performance.
Adjusted ValidClient property in player method to now not be dependant on if a client is alive.

* I didnt rename the plugin.

* ReWrite Necrophilia

Cleaned up the code structure and removed unnecesary if statement.

* ReWrite Health

* ReWrite Speed

Redesigned the whole thing sorry.

* Force Semicolon Use

* Use max level define.

Updated some older upgrades to now use a define for the max level.

* Speed Changes

Stops speed being triggered mid round and make sure  speedCoolDownTimer is triggered to reset it.

* Fix All Compile Issues But Necro

Some missing defines in necorphelia that I cannot find.

* Fixed Necro

Was missing smlib

* Update TTT Upgrades

* Add TTT Actions

Fix TTT Upgrades and changes to Player Method Map

* Fix Database Setup

Please do not look at this it will burn your retinas, as it has mine save me.

* Fix TTT Upgrades

Fixed some formating errors and wrong variables.

* Update To Pass Review

Made all requested changes.

* Update spelling in upgrades and fix for Speed.

Fixed the spelling on plugin start. Ended up with one too many c's.
Changed Plugin_Handled's to Plugin_Continue's in ttt_speed to try and fix the issue of players not moving.

* Binoculars Change.

Functions more similar to the visor in halo for the zoom where it only works for a single shot. Makes them more useful especially on larger maps.

* TTT Binoculars Fix and SQL Statement Changes

Changed some of the sql statements in player_methodmap to follow better formatting rules though want to get around to changing all these to the newer style than that off what is used now.
Set the FOV on attack in TTT Binoculars and also now play the deact sound.

* Change to new File System and Fix compatibility.

Move Upgrades to Skills.
Remove TTT Actions from Upgrades.
Add good and bad actions cookies as well as relative properties in the player methodmap.

* Delete TTT Actions

* Small Bug Fixes

Fixed some errors with prints mainly to do with Player.Msg() and some corruption in the formatting so reverted to using CPrintToChatClient.
Fixed improper brackets when calculating good action.
Added to new properties to Player for good and bad actions.
Change dev commands to new rank scheme.
Changed the fall damage hook to only be called on alive players.

* Small Changes

Renamed the define for Night Vision for consistency with the other defines.
Changed the print to chat in Scavenger to be more standardised.

* Bug Fixes

Removed redefines of TTT Roles in playermethod map so that the proper values are always used. This is as TTT Recently changed the integer values for the roles.
Removed empty lines from profile, also added profile command into TTT Upgrades.
Fixed the triggering of TTT Speed though it still triggers on one press at times.
Fixed the triggering of TTT Invisibility. Though currently issues with the player noy going invisible.

* Fix TTT Bloodlust

Changed bad kill  to work with the new ttt values. Could also be done with just &'s thanks bit flags.
Fixed incorrect comparisons on blood lust causing it to have no effect. Was exiting if the player was a traitor instead of if they are not.

* Fix TTT Speed

Clarified and fixed the float comparison for time, as it was always running  on each press as it would not trigger if the time was greater than 0.1 seconds instead of not also reset this value back to 0.5 but may lower it to 0.25.
Changed the error timeout to the standard of 2 seconds.

* Small Changes

Fix Indentation in TTT Invisibility.
Fix the comparison for ErrorTimeout in TTT Speed.

* Standardise Feedback from Upgrades.

Changed all values in strings to blue and used green and red colours for activating and deactivating respectively.

* Rename TTT Upgrades to TTT Skills

Renamed upgrades.inc to ttt_skills.inc
Renamed Player.Upgrade to Player.Skill
Changed all Skill defines from Upgrade to Skill.

* Add Safety Checks too TTT Invisibility.

All values are now reset on round end so any T will no longer be invisible, etc.

* Add Extra check on Round Start

I think this one may only be needed but I'd like to be certain just incase.

* Playtest Fixes

General Changes:
 - Fix formatting.
 - Change from using client serials too client userid's in all skills timers and in blood lust.

TTT Bloodlust:
 - Serial to userid changes.

TTT Skills:
 - Change from OnClientPutInServer too OnClientAuthorized.

TTT Health:
 - Changed to using locally defined variables where they already existed instead of refetching them.

TTT Invisibility:
 - Remove "round_start" hook and change to using TTT_OnRoundStart thus also removed HookEvents.
 - Inlined LateLoadClient and HookClient.
 - Changed to just calling the timer on Round Start and End instead of repeating code.
 - Changed TestInvis rank to Dev from Senator.

TTT Speed:
 - Changed to using CallTimer over TriggerTimer as too remove error spam(see logs).

TTT Vampire:
 - Fixed incorrect gained health message wrong formating usage now %d rather than %n.

* Unstaged Changes

* Fix upgrades compile issues.

Didnt rename Player.SetUpgrade too Player.SetSkill

* Move all functions from player methodmap

Move to only using specific parts of SMLib.
Removed the sm lib dependancy from the player methodmap as to stop broken includes.
Moved most of the functions used only in ttt skills out of the player map for tidiness.

* Remove old ttt nightvision

* Fixes

Add new natives.

* Ranks fixes

* Fixes

* Skill Speed Fix

* Fix Random Player model resetting on respawn

* Update my Ego

* Vampirism fix wrong conversion